+++ src/widgets/layout/vertical.rs
@@ -5,7 +5,7 @@ use super::{Constraints, Padding};
 use crate::widgets::ctx::{LayoutCtx, PositionCtx};
 use crate::widgets::{Direction, Expand, Spacer, WidgetContainer};
 
-pub fn layout(widgets: &mut [WidgetContainer], ctx: LayoutCtx) -> Size {
+pub fn layout(widgets: &mut [WidgetContainer], ctx: LayoutCtx, ignore_spacers_and_expansions: bool) -> Size {
     let mut used_height = 0;
     let mut width = 0;
 
@@ -30,17 +30,21 @@ pub fn layout(widgets: &mut [WidgetContainer], ctx: LayoutCtx) -> Size {
         used_height = (used_height + size.height).min(max_height);
     }
 
-    let expanded_size = expanded::layout(
-        widgets,
-        LayoutCtx::new(
-            Constraints::new(ctx.constraints.max_width, max_height - used_height),
-            ctx.force_layout,
-            Padding::ZERO,
+    let expanded_size = match ignore_spacers_and_expansions {
+        false => expanded::layout(
+            widgets,
+            LayoutCtx::new(
+                Constraints::new(ctx.constraints.max_width, max_height - used_height),
+                ctx.force_layout,
+                Padding::ZERO,
+            ),
+            Direction::Vertical,
         ),
-        Direction::Vertical,
-    );
+        true => Size::ZERO,
+    };
 
-    width = width.max(expanded_size.width).max(ctx.constraints.min_width);
+    width = width.max(expanded_size.width).max(ctx.constraints.min_width) + ctx.padding.left + ctx.padding.right;
+    width = width.min(ctx.constraints.max_width);
 
     let spacers_size = spacers::layout(
         widgets,
@@ -49,9 +53,10 @@ pub fn layout(widgets: &mut [WidgetContainer], ctx: LayoutCtx) -> Size {
         Direction::Vertical,
     );
 
-    let height = used_height + expanded_size.height + spacers_size.height;
-    let height = height.max(ctx.constraints.min_height);
-    Size::new(width, height)
+    let mut height = used_height + expanded_size.height + spacers_size.height;
+    height = height.max(ctx.constraints.min_height).min(ctx.constraints.max_height);
+
+    Size::new(width, height.min(ctx.constraints.max_height)) + ctx.padding_size()
 }
 
 pub fn position(widgets: &mut [WidgetContainer], mut ctx: PositionCtx) {