+++ CHANGES.md
@@ -1,13 +1,13 @@
 * 0.2.0
-    * `runtime::Event::scroll_up` and `runtime::Event::scroll_down` now returns
-      the screen position and any key modifiers.
+    * `ScrollView` widget was removed
     * `DataCtx` now has `get_ref` counter part to `get_mut`.
     * `Border` has an optional `min-width` and `min-height`.
     * `VStack` / `HStack` and `ZStack` now have an optional `min-width` and
       `min-height`.
     * Convenient `String` access on the `Text` widget via
       `get_text_mut(span_index)` and `get_text(span_index)`.
-    * BUGFIX: Padding for `ZStack` and `Alignment` is now working.
+    * BUGFIX: Padding for `ZStack`, `Alignment` and `Expanded` is now working.
+    * Rename feature "with-flume" to "flume"
 * 0.1.2
     * `WidgetContainer::by_id` is made more ergonomic and can now called with a
       string slice now.